Framework for constraints (not finished yet).

+++ FelzenszwalbSegmentation/convolve.h
@@ -31,7 +31,7 @@ static void convolve_even(image<float> *src, image<float> *dst,
 			  std::vector<float> &mask) {
   int width = src->width();
   int height = src->height();
-  int len = mask.size();
+  int len = (int)mask.size();
 
   for (int y = 0; y < height; y++) {
     for (int x = 0; x < width; x++) {
@@ -51,7 +51,7 @@ static void convolve_odd(image<float> *src, image<float> *dst,
 			 std::vector<float> &mask) {
   int width = src->width();
   int height = src->height();
-  int len = mask.size();
+  int len = (int)mask.size();
 
   for (int y = 0; y < height; y++) {
     for (int x = 0; x < width; x++) {
